DOM elements, buttons, css
Buttons are touched up but their positioning needs a little work

Canvases are now created in javascript to be more flexible. They should
be created using the parent DOM size and position as a reference though
- look into passing that into init.

Joystick needs to have a settable size and give some kind of feedback
that the user can use.

CSS was moved to javascript for flexibility for the user.